vmware fixes

+++ contrail_setup_utils/compute-server-setup.sh
@@ -59,12 +59,13 @@ if [ $CONTROLLER != $COMPUTE ] ; then
     openstack-config --set /etc/nova/nova.conf keystone_authtoken auth_host $CONTROLLER
 fi
 
-if [ -f /etc/nova/nova-compute.conf ]; then
+if [ $VMWARE_IP ]; then
     openstack-config --del /etc/nova/nova.conf DEFAULT compute_driver
     openstack-config --set /etc/nova/nova.conf DEFAULT compute_driver vmwareapi.ContrailESXDriver
-
-    openstack-config --del /etc/nova/nova-compute.conf DEFAULT compute_driver
-    openstack-config --set /etc/nova/nova-compute.conf DEFAULT compute_driver vmwareapi.ContrailESXDriver
+    if [ -f /etc/nova/nova-compute.conf ]; then
+        openstack-config --del /etc/nova/nova-compute.conf DEFAULT compute_driver
+        openstack-config --set /etc/nova/nova-compute.conf DEFAULT compute_driver vmwareapi.ContrailESXDriver
+    fi
 fi
 
 openstack-config --set /etc/nova/nova.conf DEFAULT ec2_private_dns_show_ip False
